Refactor refresh token logic into a service
